Updated code to separate classes into their own files
